warning fix
git-svn-id: http://miranda.googlecode.com/svn/trunk@9709 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/IRCG/services.cpp
@@ -1147,7 +1147,7 @@ void CIrcProto::DisconnectFromServer(void)
 
 INT_PTR __cdecl CIrcProto::GetMyAwayMsg(WPARAM wParam,LPARAM lParam)
 {
-	if ( wParam != m_iStatus )
+	if (( int )wParam != m_iStatus )
 		return 0;
 
 	const TCHAR* p = m_statusMessage.c_str();
